+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 🚩 Deprecations 🚩
 
 - Deprecate `p[metric|log|trace]otlp.RegiserServer` in favor of `p[metric|log|trace]otlp.RegiserGRPCServer` (#6180)
+- Deprecate `pcommon.NewValueString` in favor of `pcommon.NewValueStr` (#6209)
 
 ### 💡 Enhancements 💡